+++ www/src/pages/contributing/__tests__/stub-list.js
@@ -32,7 +32,9 @@ jest.mock(`react-modal`, () => {
 import React from "react"
 import { useStaticQuery } from "gatsby"
 import { render } from "@testing-library/react"
+import { ThemeProvider } from "theme-ui"
 
+import theme from "../../../../src/gatsby-plugin-theme-ui"
 import StubList from "../stub-list"
 
 let location
@@ -51,14 +53,22 @@ beforeEach(() => {
 
 describe(`StubList`, () => {
   it(`shows a title`, () => {
-    const { getByText } = render(<StubList location={location} />)
+    const { getByText } = render(
+      <ThemeProvider theme={theme}>
+        <StubList location={location} />
+      </ThemeProvider>
+    )
     const element = getByText(`Stub List`)
 
     expect(element).toBeVisible()
   })
 
   it(`shows a call to action to edit stubs`, () => {
-    const { getByText } = render(<StubList location={location} />)
+    const { getByText } = render(
+      <ThemeProvider theme={theme}>
+        <StubList location={location} />
+      </ThemeProvider>
+    )
 
     expect(getByText(`How to Write a Stub`)).toBeVisible()
   })
@@ -66,7 +76,11 @@ describe(`StubList`, () => {
   describe(`stubs`, () => {
     let stubs
     beforeEach(() => {
-      const { getByTestId } = render(<StubList location={location} />)
+      const { getByTestId } = render(
+        <ThemeProvider theme={theme}>
+          <StubList location={location} />
+        </ThemeProvider>
+      )
       stubs = getByTestId(`list-of-stubs`).querySelectorAll(`li`)
     })